requirements.txt
django
psycopg2-binary
djangorestframework
django-rest-passwordreset
djangorestframework-yaml
